Убрал Owner и Client

+++ src/main/java/org/example/service/Impl/UserServiceImpl.java
@@ -1,9 +1,10 @@
 package org.example.service.Impl;
 
 import org.example.dao.*;
-import org.example.enums.RoleType;
 import org.example.repository.PersonRepository;
 import org.example.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.*;
 import org.springframework.stereotype.Service;
@@ -15,12 +16,14 @@ import java.util.List;
 @Transactional
 public class UserServiceImpl implements UserDetailsService, UserService {
 
+    private static final Logger log = LoggerFactory.getLogger(UserServiceImpl.class);
     private final PersonRepository personRepository;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return personRepository.findByLogin(username).orElseThrow(() -> new UsernameNotFoundException(username));
     }
+
     @Autowired
     public UserServiceImpl(PersonRepository personRepository) {
         this.personRepository = personRepository;
@@ -29,17 +32,32 @@ public class UserServiceImpl implements UserDetailsService, UserService {
 
     @Override
     public Person createUser(Person person) {
-        return personRepository.save(person);
+        Person personByEmail = personRepository.findByEmail(person.getEmail()).orElse(null);
+        if (personByEmail==null) {
+            return personRepository.save(person);
+        }
+        log.error("Пользователь с таким email уже существует");
+        return personByEmail;
     }
 
     @Override
     public Person updateUser(Person person) {
+        if (personRepository.findByLogin(person.getLogin()).isPresent() || personRepository.findByEmail(person.getEmail()).isPresent()) {
+            return null;
+        }
         return personRepository.save(person);
     }
 
     @Override
     public void deleteUserById(Long id) {
-        personRepository.deleteById(id);
+        if (!personRepository.findById(id).isPresent()) {
+            log.info("not deleting user with id: " + id);
+            throw new UsernameNotFoundException("User not found");
+        } else {
+            personRepository.deleteById(id);
+            log.info("deleting user with id: " + id);
+        }
+
     }
 
     @Override
@@ -49,7 +67,7 @@ public class UserServiceImpl implements UserDetailsService, UserService {
 
     @Override
     public Person getUserById(long id) {
-        return  personRepository.findById(id).orElse(null);
+        return personRepository.findById(id).orElse(null);
     }
 
     @Override
